// freelancing-management-system/src/app/models/user.model.ts
export interface User {
 id?: number;
  fullName? : string;
  country? : string;
  hourlyRate? : number;
  languages? : string;
  education? : string;
  certifications? : string;
  name?: string;
  email?: string;
  password?: string;
  role?: string;
  bio?: string;
  skills?: string;
  experience?: string;
  avatar?: string;
}
// private String fullName;
// private String title;
// private String country;
// private Double hourlyRate;
// private String skills;
// private String languages;
// private String education;
// private String certifications;
// private Integer experience;
// private String bio;